requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
langchain-chroma
langchain-groq
python-dotenv
arize-phoenix
openai